added GANRunner to catalyst.dl.runner
