+++ catalyst/dl/runner/__init__.py
@@ -1,6 +1,6 @@
 # flake8: noqa
 
-from .gan import GanRunner
+from .gan import MultiPhaseRunner, GanRunner
 from .supervised import SupervisedRunner
 
 from catalyst.contrib.runner import *  # isort:skip
