+++ examples/mnist_gan/config.yml
@@ -1,31 +1,34 @@
+common_hyperparams:
+  noise_dim: &noise_dim 16
+  # ensure the names below are the same as in your runner ("discriminator_train" & "generator_train")
+  discriminator_train_phase: &discriminator_train_phase discriminator_train
+  generator_train_phase: &generator_train_phase generator_train
+
 model_params:
   _key_value: True
   generator:
     model: SimpleGenerator
-    noise_dim: &noise_dim 16
+    noise_dim: *noise_dim
   discriminator:
     model: SimpleDiscriminator
 
 
 args:
   expdir: "mnist_gan"
-  logdir: "./logs/mnist_gan"
+  baselogdir: "./logs/mnist_advanced_gans/vanilla_gan"
 
 
 stages:
 
   data_params:
     batch_size: 64
-    num_workers: 1
+    num_workers: 0
 
   state_params:
     num_epochs: 100
     main_metric: &reduce_metric loss_g
     minimize_metric: True
     batch_consistant_metrics: False
-    noise_dim: *noise_dim
-    discriminator_train_phase: &discriminator_train_phase discriminator_train
-    generator_train_phase: &generator_train_phase generator_train
 
   criterion_params:
     criterion: BCEWithLogitsLoss
@@ -38,6 +41,13 @@ stages:
       train_phases:
         *discriminator_train_phase: 1
         *generator_train_phase: 1
+
+    prepare_batch_noise:
+      callback: AddBatchNoiseCallback
+      noise_shape: *noise_dim
+
+    prepare_batch_targets:
+      callback: AddRealFakeTargetsCallback
 
     loss_g:
       _wrapper: &g_train_wrapper
@@ -82,7 +92,7 @@ stages:
 
     visualizer:
       callback: VisualizationCallback
-      output_keys: "fake_features"
+      output_keys: "fake_data"
       n_row: 5
       max_images: 25
